Add files via upload
🔐 Python Mini-Project: Password Generator 🔐

Welcome to my latest Python project video! In this tutorial, we’ll be creating a secure and customizable password generator using Python. Whether you’re a beginner or looking to enhance your coding skills, this project is perfect for you. 😊

📜 What You'll Learn:
How to use the random module to generate random characters
How to work with lists and loops in Python
How to enhance password security by shuffling characters
📌 Project Features:
Easy Level:
Sequentially generates a password with user-specified numbers of letters, symbols, and numbers.
Hard Level:
Shuffles the characters to create a more secure password.

👍 Don't Forget to:
Like the video if you found it helpful 👍
Subscribe to my channel for more awesome Python projects and tutorials 🔔
Share your feedback and questions in the comments section below 💬


Happy coding! 🚀✨